Create README.md
Unleash the power of concise links with our URL Shortener! Crafted with Flask and fortified by SQLite, this project not only trims your URLs but ensures top-tier security with hashed passwords. From QR code magic to user-specific link management, it's the gateway to streamlined navigation in the vast expanse of the web. Your URLs, your way, with a touch of elegance!